refactor(cli): give the preview, the scoped root and the S3 stubs one home each
Follow-up pass over the remote-data-root work. No behaviour change; the
suite is unchanged at 2110 passed.

DataRoot.scoped replaces the delegating wrapper that had grown in
deltatables.py only because the agent writing it could not reach
data_root.py. Both implementations now return a real root of their own
class: a local sub-directory is just a directory, and the S3 view
re-indexes the objects the parent already listed, so a forty-run prefix
still costs one listing rather than forty-one. Being real roots, they have
runs() instead of conspicuously lacking it.

The preview moves out to template_preview.py. templates.py had grown to
1744 lines and the cut is clean in one direction: the preview reads
resolve_template, resolution never reads the preview.

The preview status vocabulary was a bare string in the producer and a
parallel tuple in the renderer. It is now one Literal, declared where the
statuses are assigned and pointed at from run.py.

Four S3 test stubs became one, in a non-test module so pytest does not
collect it. Test modules were importing helpers from other test modules,
which is its own smell. Every file keeps its exact test count.

Two things deliberately left as they are, now with comments saying why.
The glob-to-regex translation in data_root.py and the fnmatch matching in
scan.py look like duplicates but hold opposite contracts, on whether a
star crosses a slash, and each is right for its caller. as_data_root and
recipes._as_data_root likewise differ: one coerces a user-supplied
location and understands s3://, the other adapts a recipe data_dir that
has always meant a directory on disk.
