+++ depictio/cli/cli/utils/templates.py
@@ -20,7 +20,6 @@ from __future__ import annotations
 import copy
 import json
 import re
-from dataclasses import dataclass, field
 from datetime import datetime
 from pathlib import Path
 from typing import Any
@@ -28,7 +27,7 @@ from typing import Any
 import httpx
 import yaml
 
-from depictio.cli.cli.utils.data_root import DataRoot, data_root_for
+from depictio.cli.cli.utils.data_root import DataRoot, as_data_root
 from depictio.cli.cli_logging import logger
 from depictio.models.models.templates import (
     DCOverride,
@@ -46,31 +45,8 @@ _TEMPLATE_VAR_RE = re.compile(r"\{([A-Z0-9_]+)\}")
 
 # Recorded on the expected-DC manifest for a DC dropped by
 # `_prune_missing_optional_single_file_dcs`, and read back by
-# `preview_data_root` to tell that pruning apart from conditional gating.
+# `template_preview` to tell that pruning apart from conditional gating.
 OPTIONAL_SOURCE_MISSING_REASON = "optional source file not found"
-
-
-def as_data_root(value: str | Path | DataRoot | None, CLI_config=None) -> DataRoot | None:
-    """A :class:`DataRoot` for a location string, a ``Path``, or an existing root.
-
-    ``None`` stays ``None`` (manifest-driven templates have no data root at
-    all). A local path is made absolute first, so the root's own ``location``
-    is what ``{DATA_ROOT}`` substitutes to; a URL is passed through verbatim,
-    since ``Path("s3://b/k").absolute()`` would silently produce
-    ``<cwd>/s3:/b/k`` and corrupt every path derived from it.
-
-    Accepting an already-built root is what lets a caller that needs the root
-    itself (``preview_data_root``) hand the same one to ``resolve_template``
-    instead of paying for a second S3 listing.
-    """
-    if value is None:
-        return None
-    if not isinstance(value, (str, Path)):
-        return value
-    location = str(value)
-    if "://" not in location:
-        location = str(Path(location).absolute())
-    return data_root_for(location, CLI_config)
 
 
 # A template version directory: purely numeric dotted segments (e.g. "2.16.0").
@@ -1098,6 +1074,7 @@ def resolve_template(
     5. Substitutes template variables in all paths
     6. Applies conditional rules (remove DCs, prune links, select dashboards)
     6a. Repoints every DC at a remote root, when the root is one
+    6b. Prunes optional single-file DCs whose source is absent
     6c. Materializes recipe DCs that ship a pre-computed seed
     7. Strips hardcoded IDs
     8. Sets project name
@@ -1435,310 +1412,3 @@ def import_dashboards_from_template(
         results.append(entry)
 
     return results
-
-
-# ---------------------------------------------------------------------------
-# Data-root preview
-# ---------------------------------------------------------------------------
-
-# The variables worth showing before a run: the ones resolution *derived* rather
-# than the ones the user typed. The SKIP_*/IS_* flags matter most, because they
-# are what the conditionals gate on, so they decide which DCs exist at all.
-_PREVIEW_VARIABLE_NAMES = (
-    "SAMPLESHEET_FILE",
-    "METADATA_FILE",
-    "METADATA_ID_COL",
-    "GROUP_COL",
-    "GROUP_COL_DISPLAY",
-    "ANNOTATION_COLS",
-)
-_PREVIEW_FLAG_PREFIXES = ("SKIP_", "IS_")
-
-
-@dataclass
-class DataCollectionPreview:
-    """What one data collection would find under a given data root.
-
-    ``matched`` is a count taken with the rule the scan itself will use, so an
-    ``empty`` row means the scan will genuinely find nothing, not that the
-    preview looked somewhere else. Where nothing can be counted from the root -
-    a manifest that is fetched at ingest, a URL on another host - the row is
-    ``ok`` with ``matched`` 0 and ``location`` says where it points.
-    """
-
-    tag: str
-    kind: str  # "scan" | "recipe"
-    mode: str | None  # scan mode, None for recipe DCs
-    location: str  # what it looked at, human readable
-    matched: int  # files found
-    missing_sources: list[str] = field(default_factory=list)  # unresolvable recipe sources
-    optional: bool = False
-    status: str = "ok"  # "ok" | "empty" | "missing" | "pruned"
-
-
-@dataclass
-class RunPreview:
-    """What ``--data-root <location>`` would actually yield, before anything is created."""
-
-    template_id: str
-    data_root: str
-    project_name: str
-    resolved_variables: dict[str, str]
-    detected_runs: list[str]
-    data_collections: list[DataCollectionPreview]
-    pruned_optional_dcs: list[str] = field(default_factory=list)
-    dashboards: list[str] = field(default_factory=list)
-    truncated: bool = False  # the listing hit its key cap, so the view is partial
-
-
-def _scan_match_regex(mode: str, parameters: dict) -> str:
-    """The regex a pattern-based scan will match files with.
-
-    ``recursive`` carries a regex already (with the template's wildcards
-    expanded exactly as the walk expands them); ``s3_prefix`` carries a glob
-    unless it says otherwise, so translate it to the one dialect the data root's
-    matcher speaks.
-    """
-    from fnmatch import translate
-
-    from depictio.cli.cli.utils.scan_utils import construct_full_regex
-    from depictio.models.models.data_collections import Regex
-
-    if mode == "recursive":
-        regex_config = parameters.get("regex_config") or {}
-        return construct_full_regex(
-            Regex(
-                pattern=regex_config.get("pattern") or ".*",
-                wildcards=regex_config.get("wildcards"),
-            )
-        )
-    pattern = parameters.get("pattern") or "*"
-    if str(parameters.get("pattern_syntax") or "glob").lower() == "regex":
-        return pattern
-    return translate(pattern)
-
-
-def _preview_scan_dc(
-    tag: str,
-    dc_config: dict,
-    root: DataRoot,
-    runs: list[str],
-    optional: bool,
-) -> DataCollectionPreview:
-    """One row for a DC that acquires its data by scanning."""
-    scan = dc_config.get("scan") or {}
-    mode = str(scan.get("mode") or "").lower()
-    parameters = scan.get("scan_parameters") or {}
-
-    if mode in ("single", "url"):
-        location = parameters.get("filename") or parameters.get("url") or ""
-        relative = root.relative_of(location) if location else None
-        if relative is None:
-            # Not under this root (a foreign https:// URL, or a path elsewhere
-            # on disk). We cannot count it, and must not report it missing.
-            return DataCollectionPreview(
-                tag=tag, kind="scan", mode=mode, location=location, matched=0, optional=optional
-            )
-        found = root.exists(relative)
-        return DataCollectionPreview(
-            tag=tag,
-            kind="scan",
-            mode=mode,
-            location=root.url(relative),
-            matched=1 if found else 0,
-            optional=optional,
-            status="ok" if found else "missing",
-        )
-
-    if mode == "manifest":
-        # The manifest is fetched at ingest, not here: its entries are not
-        # visible from the data root's listing.
-        return DataCollectionPreview(
-            tag=tag,
-            kind="scan",
-            mode=mode,
-            location=parameters.get("manifest_url") or "",
-            matched=0,
-            optional=optional,
-        )
-
-    if mode in ("recursive", "s3_prefix"):
-        regex = _scan_match_regex(mode, parameters)
-        within = ""
-        location = root.url("")
-        if mode == "s3_prefix":
-            prefix = parameters.get("prefix") or ""
-            location = prefix
-            relative = root.relative_of(prefix)
-            if relative is None:
-                # A prefix on another bucket: outside what this root listed.
-                return DataCollectionPreview(
-                    tag=tag,
-                    kind="scan",
-                    mode=mode,
-                    location=prefix,
-                    matched=0,
-                    optional=optional,
-                )
-            within = relative
-        matched = (
-            sum(len(root.match(regex, within=f"{run}/{within}".strip("/"))) for run in runs)
-            if runs
-            else len(root.match(regex, within=within))
-        )
-        return DataCollectionPreview(
-            tag=tag,
-            kind="scan",
-            mode=mode,
-            location=location,
-            matched=matched,
-            optional=optional,
-            status="ok" if matched else "empty",
-        )
-
-    return DataCollectionPreview(
-        tag=tag, kind="scan", mode=mode or None, location=root.url(""), matched=0, optional=optional
-    )
-
-
-def _override_binding(override: Any, attribute: str) -> str | None:
-    """One field of a source override, whether it arrived as a dict or a model."""
-    if isinstance(override, dict):
-        return override.get(attribute)
-    return getattr(override, attribute, None)
-
-
-def _preview_recipe_dc(
-    tag: str,
-    dc_config: dict,
-    root: DataRoot,
-    optional: bool,
-) -> DataCollectionPreview:
-    """One row for a ``source: transformed`` DC, resolved through its recipe's SOURCES.
-
-    ``dc_ref`` sources are skipped: they are satisfied from another DC's output,
-    not from a file under the root, so their absence here says nothing.
-    """
-    transform = dc_config.get("transform") or {}
-    recipe_name = transform.get("recipe") or ""
-    row = DataCollectionPreview(
-        tag=tag, kind="recipe", mode=None, location=root.url(""), matched=0, optional=optional
-    )
-    if not recipe_name:
-        return row
-
-    try:
-        from depictio.recipes import load_recipe
-
-        module = load_recipe(recipe_name)
-    except Exception as exc:  # noqa: BLE001 - a preview never fails the run
-        logger.warning(f"Preview: could not load recipe '{recipe_name}' for '{tag}': {exc}")
-        return row
-
-    overrides = transform.get("source_overrides") or {}
-    for source in module.SOURCES:
-        if source.dc_ref is not None:
-            continue
-        override = overrides.get(source.ref)
-        glob_pattern = _override_binding(override, "glob_pattern") if override else None
-        path = _override_binding(override, "path") if override else None
-        if glob_pattern is None and path is None:
-            glob_pattern, path = source.glob_pattern, source.path
-
-        if glob_pattern:
-            hits = len(root.glob(glob_pattern))
-        elif path:
-            hits = 1 if root.exists(path) else 0
-        else:
-            continue
-        row.matched += hits
-        if not hits and not source.optional:
-            row.missing_sources.append(glob_pattern or path or source.ref)
-
-    row.status = "missing" if row.missing_sources else "ok"
-    return row
-
-
-def preview_data_root(
-    template_id: str,
-    data_root: str,
-    variables: dict[str, str] | None = None,
-    CLI_config=None,
-) -> RunPreview:
-    """What a template would resolve to against a data root, without creating anything.
-
-    Answers the question a user actually has before running an ingestion: given
-    this directory or this ``s3://`` prefix, which data collections will find
-    their data, which will come up empty, and which variables did the run's own
-    parameters decide. It resolves the template exactly as ``depictio run``
-    would - same auto-detection, same conditionals, same pruning - and then
-    counts each surviving DC's matches with the rule its scan will use.
-
-    Every question is asked of the same :class:`DataRoot`, so a remote prefix
-    costs one listing for the whole preview.
-
-    Args:
-        template_id: Template identifier, or a path to a template bundle.
-        data_root: The directory or ``s3://`` prefix to preview.
-        variables: ``--var`` values, exactly as ``resolve_template`` takes them.
-        CLI_config: Used to build a remote root's S3 client.
-    """
-    root = as_data_root(data_root, CLI_config)
-    if root is None:
-        raise ValueError("preview_data_root needs a data root; got None")
-
-    config, template_metadata, template_origin, dashboard_paths, resolved = resolve_template(
-        template_id=template_id,
-        data_root=root,
-        extra_vars=dict(variables) if variables else None,
-        CLI_config=CLI_config,
-    )
-
-    rows: list[DataCollectionPreview] = []
-    detected_runs: list[str] = []
-    for workflow in config.get("workflows") or []:
-        data_location = workflow.get("data_location") or {}
-        runs: list[str] = []
-        if data_location.get("structure") == "sequencing-runs" and data_location.get("runs_regex"):
-            # One run directory per scan pass, the same way the walk scopes it.
-            runs = root.runs(data_location["runs_regex"])
-            detected_runs.extend(run for run in runs if run not in detected_runs)
-
-        for dc in workflow.get("data_collections") or []:
-            tag = dc.get("data_collection_tag") or ""
-            dc_config = dc.get("config") or {}
-            optional = bool(dc.get("optional"))
-            # A materialized recipe DC has a scan block over its seed, so it is
-            # previewed as what it now is: a file scan.
-            if dc_config.get("source") == "transformed" and not dc_config.get("scan"):
-                rows.append(_preview_recipe_dc(tag, dc_config, root, optional))
-            else:
-                rows.append(_preview_scan_dc(tag, dc_config, root, runs, optional))
-
-    pruned = [
-        entry.data_collection_tag
-        for entry in template_origin.expected_data_collections
-        if not entry.included and entry.removal_reason == OPTIONAL_SOURCE_MISSING_REASON
-    ]
-    rows.extend(
-        DataCollectionPreview(
-            tag=tag, kind="scan", mode=None, location="", matched=0, optional=True, status="pruned"
-        )
-        for tag in pruned
-    )
-
-    return RunPreview(
-        template_id=template_metadata.template_id,
-        data_root=root.location,
-        project_name=config.get("name") or "",
-        resolved_variables={
-            name: value
-            for name, value in resolved.items()
-            if name in _PREVIEW_VARIABLE_NAMES or name.startswith(_PREVIEW_FLAG_PREFIXES)
-        },
-        detected_runs=detected_runs,
-        data_collections=rows,
-        pruned_optional_dcs=pruned,
-        dashboards=[path.name for path in dashboard_paths],
-        truncated=bool(getattr(root, "truncated", False)),
-    )
